Refactor RF4SConfigBridge
Refactor RF4SConfigBridge.ts into smaller, more manageable files, preserving all existing functionality.
